+++ packages/narrat/src/utils/images-loader.ts
@@ -3,74 +3,73 @@ import { error } from './error-handling';
 import { logger } from './logger';
 
 export const images: {
-  [key: string]: HTMLImageElement;
+  [key: string]: HTMLImageElement | Promise<HTMLImageElement>;
 } = {};
-let imagesToLoad = 0;
-let imagesLoaded = 0;
 
-export function loadImages(config: Config): Promise<void> {
+export async function getImage(path: string): Promise<HTMLImageElement> {
+  if (images[path]) {
+    return images[path];
+  }
+  return downloadImage(path, path);
+}
+
+export async function loadImages(config: Config): Promise<HTMLImageElement[]> {
   logger.log(`Loading images`);
-  return new Promise((resolve, reject) => {
-    if (Object.keys(config.images).length < 1) {
-      resolve();
-      return;
+  let loadedCount = 0;
+  let toLoadCount = 0;
+  const allImages: Array<Promise<HTMLImageElement>> = [];
+  const loadImage = async (key: string, path: string) => {
+    toLoadCount++;
+    const image = await downloadImage(key, path);
+    loadedCount++;
+    console.log(`Loaded ${loadedCount} / ${toLoadCount} images`);
+    return image;
+  };
+  const addImage = (key: string, path: string) => {
+    allImages.push(loadImage(key, path));
+  };
+  for (const key in config.images) {
+    const path = config.images[key];
+    addImage(key, path);
+  }
+  // Find other images to load from config
+  for (const key in config.screens) {
+    const screen = config.screens[key];
+    if (screen.background && !config.images[screen.background]) {
+      addImage(screen.background, screen.background);
     }
-    for (const key in config.images) {
-      const path = config.images[key];
-      loadImage(key, path, resolve, reject);
-    }
-    // Find other images to load from config
-    for (const key in config.screens) {
-      const screen = config.screens[key];
-      if (screen.background && !config.images[screen.background]) {
-        loadImage(screen.background, screen.background, resolve, reject);
-      }
-      if (screen.buttons) {
-        for (const buttonKey in screen.buttons) {
-          const button = screen.buttons[buttonKey];
-          if (typeof button === 'object') {
-            if (button.background && !config.images[button.background]) {
-              loadImage(button.background, button.background, resolve, reject);
-            }
+    if (screen.buttons) {
+      for (const buttonKey in screen.buttons) {
+        const button = screen.buttons[buttonKey];
+        if (typeof button === 'object') {
+          if (button.background && !config.images[button.background]) {
+            addImage(button.background, button.background);
           }
         }
       }
     }
-    for (const key in config.buttons) {
-      const button = config.buttons[key];
-      if (typeof button === 'object') {
-        if (button.background && !config.images[button.background]) {
-          loadImage(button.background, button.background, resolve, reject);
-        }
-      }
-    }
-  });
+  }
+  return Promise.all(allImages);
 }
 
-export function loadImage(
-  key: string,
-  path: string,
-  resolver: any,
-  rejecter: any,
-) {
+export async function downloadImage(key: string, path: string) {
   if (images[key]) {
-    return;
+    return images[key];
   }
-  imagesToLoad++;
-  logger.log(`Loading image ${key} at ${path}`);
-  const image = new Image();
-  image.onload = () => {
-    imagesLoaded += 1;
-    images[key] = image;
-    logger.log(`Loaded image ${key} successfully`);
-    if (imagesLoaded >= imagesToLoad) {
-      logger.log(`All images loaded`);
-      resolver();
-    }
-  };
-  image.onerror = (e) => {
-    error(`Failed to load image ${key} at ${path}`, e);
-    rejecter(e);
-  };
-  image.src = getAssetUrl(path);
+  const promise = new Promise<HTMLImageElement>((resolve, reject) => {
+    logger.log(`Loading image ${key} at ${path}`);
+    const image = new Image();
+    image.onload = () => {
+      images[key] = image;
+      logger.log(`Loaded image ${key} successfully`);
+      resolve(image);
+    };
+    image.onerror = (e) => {
+      error(`Failed to load image ${key} at ${path}`, e);
+      reject(e);
+    };
+    image.src = getAssetUrl(path);
+  });
+  images[key] = promise;
+  return promise;
 }